Implement Walking (#5)

+++ src/component/transform.rs
@@ -1,5 +1,5 @@
 #[derive(Debug, Clone)]
 pub struct Transform {
-    pub x: f32,
-    pub y: f32
+    pub x: f64,
+    pub y: f64
 }
